pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-scaffolding-multipledatabases"
version = "0.1.0"
description = "FastAPI server with multiple database support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.22",
    "asyncpg>=0.28.0",  # PostgreSQL async driver
    "aiomysql>=0.2.0",  # MySQL async driver
    "motor>=3.3.1",     # MongoDB async driver
    "polars>=0.19.12",  # Preferred over pandas
    "httpx>=0.25.0",    # For microservice communication
    "tenacity>=8.2.3",  # For retries
    "structlog>=23.2.0", # Structured logging
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.6.1",
    "ruff>=0.0.292",
]

sqlserver = [
    "pyodbc>=4.0.39",  # SQL Server driver
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]
ignore = []

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
packages = ["app"]
